+++ app/google-services.json
@@ -31,6 +31,43 @@
           }
         },
         {
+          "client_id": "740114569992-tj7ostdigefvu1nbqcgic1hlrbnk8ugb.apps.googleusercontent.com",
+          "client_type": 1,
+          "android_info": {
+            "package_name": "com.example.mobile_programing",
+            "certificate_hash": "041bc366062f8a6d81c11dc70579fe96bcddf7d4"
+          }
+        },
+        {
+          "client_id": "740114569992-f1io8l0m6c6t24k9fo5s2b2ohdc4psa5.apps.googleusercontent.com",
+          "client_type": 3
+        }
+      ],
+      "api_key": [
+        {
+          "current_key": "AIzaSyAw40MgMYVnxKF6yYkyUigUeoH_D8ZnBjs"
+        }
+      ],
+      "services": {
+        "appinvite_service": {
+          "other_platform_oauth_client": [
+            {
+              "client_id": "740114569992-f1io8l0m6c6t24k9fo5s2b2ohdc4psa5.apps.googleusercontent.com",
+              "client_type": 3
+            }
+          ]
+        }
+      }
+    },
+    {
+      "client_info": {
+        "mobilesdk_app_id": "1:740114569992:android:fcd5d5487d910a58546abd",
+        "android_client_info": {
+          "package_name": "com.routine_mate.mobile_programing"
+        }
+      },
+      "oauth_client": [
+        {
           "client_id": "740114569992-f1io8l0m6c6t24k9fo5s2b2ohdc4psa5.apps.googleusercontent.com",
           "client_type": 3
         }
