--- app/src/main/java/com/example/mobile_programing/models/Sample.kt
@@ -1,8 +1,0 @@
-package com.example.mobile_programing.models
-
-// Model 파일 sample
-class Sample(val index: Int, val name: String) {
-
-}
-
-// Model 파일 sample
