+++ app/src/main/java/com/routine_mate/mobile_programing/models/Sample.kt
@@ -1,0 +1,8 @@
+package com.routine_mate.mobile_programing.models
+
+// Model 파일 sample
+class Sample(val index: Int, val name: String) {
+
+}
+
+// Model 파일 sample
