+++ app/src/main/res/layout/activity_routine_progress.xml
@@ -6,7 +6,7 @@
     <data>
         <variable
             name="routineProgressViewModel"
-            type="com.example.mobile_programing.viewModel.RoutineProgressViewModel" />
+            type="com.routine_mate.mobile_programing.viewModel.RoutineProgressViewModel" />
     </data>
 <androidx.constraintlayout.widget.ConstraintLayout
     android:id="@+id/cl_routine_progress"
@@ -57,7 +57,7 @@
 
     <androidx.fragment.app.FragmentContainerView
         android:id="@+id/fl_routine_progress"
-        android:name="com.example.mobile_programing.views.fragments.CardDisplayFragment"
+        android:name="com.routine_mate.mobile_programing.views.fragments.CardDisplayFragment"
         android:layout_width="0dp"
         android:layout_height="0dp"
         app:layout_constraintBottom_toTopOf="@+id/et_card_memo"
